Safari-609.1.11.1/LayoutTests/http/tests/websocket/tests/hybi/contentextensions/display-none-worker-expected.txt

CONSOLE MESSAGE: line 23: worker onopen
CONSOLE MESSAGE: line 23: worker onmessage reply sent
CONSOLE MESSAGE: line 23: PASS worker onclose
This should not be hidden.
